Add `cd` function to remind me what I was doing
To reduce the friction of working on projects, I have a
`.quick-commence.txt` file in my project directories. It's usually just
a few answers to the questions of "What was I doing?" and "What's the
next step?".

Having a reminder of these things when I change into a project
directory helps me get immersed into the project a lot faster.